backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
psycopg[binary]>=3.2.10
sqlalchemy==2.0.38
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
google-generativeai==0.8.4
pinecone-client==5.0.1
tavily-python==0.5.0
bcrypt==4.3.0
requests==2.32.3
beautifulsoup4==4.13.3
lxml==5.3.1
sympy==1.13.3
httpx==0.28.1
